[FIX] web_studio: allow modification of multi-occurrence fields
Before this commit, modifying the attributes of a field that appears
multiple times in a single view may not work.

Whether or not it would work depended on whether it was the first
element of the view or not.

E.g.: If clicking on the second occurrence of the view and changing the
name, the change would be applied to the first occurrence instead.

This happens because the client sends the server the minimum amount of
information possible whenever handling a field in studio, if the field
has an identifying attribute (such as name) we just send the tag and
the attribute and the server will have to "make do" with just that
information (relative xpath), if the target node does not have
identifying attributes then it will send an `xpath_info` object with
which the server can build an absolute xpath.

With this commit, changing the attributes of any field with multiple
occurrences within a view, whether they're first, last or in the middle,
modifies the attributes of the correct node and not the ones of the
first.

This is achieved by a modification in the client and in the server:
    - The client will *always* send `xpath_info` to the server
    - The server will *always* use the absolute xpath first and will
    only use relative ones if there's no other choice

Basing view modifications on absolute xpaths as much as possible
guarantees that the target element is most likely the correct one
whether there's a single one or a hundred with the same attributes.

closes odoo/enterprise#11104

Task: 2227691
X-original-commit: 6d41cc3223bb98868214284a3e5a04874c8b14dc
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>
Signed-off-by: Adrian Torres (adt) <adt@odoo.com>