[FIX] quality_control: transfer quality check to backorder
This commit makes it so a picking that contains quality checks transfers
incomplete quality checks to any backorder pickings it creates. Note
that this means a partially completed move of a product with a quality
check will not have its quality check moved to the backorder since the
check will no longer be "To Do" at this point.

To reproduce bug:
1. Create a picking with 2 products, where at least 1 product has a control
   point in that picking operation
2. Validate the picking without the product that has a control point in
   that picking operation and create a backorder.
3. Open backorder and it will not have a quality check for the product even
   though it should.

Bug noticed while working on inventory usability task.
Task: 2206279
Related COM PR: odoo/odoo#51681