[FIX] web_mobile: inconsistent image field in contact sync
When the user attempts to add a contact to its device's address book,
either the app crashes (iOS) or silently ignore the contact's image.

This is due to the rename of the partner's field "image" to "image_1920"
performed in commit odoo/enterprise@fabd84c1fedde92b23f2cd136cd299e20870dc06 which created
an inconsistency between the usage of the native `addContact` method and
its implementation in the mobile apps (on both iOS & Android).

This commit fixes it by mapping the new field's name to the old one to
keep descending compatibility with version < 13.0.

Also, it removes the `title` field's retrieval as it isn't used on the
mobile apps.

opw-2232760

closes odoo/enterprise#10730

X-original-commit: b67240dabfd7a238c93d46fd6ede4ab86f9d04ed
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>