[IMP] project: Improve global UX (back2basics)
Improve the onboarding of new users and clean a few screens.

After this commit, adding the attribute hide_title="true" to a map tag
will hide the panel title.

The purpose is to hide the title in the map view of the project.task
model when the project is in the context.

Task ID: 2237996

closes odoo/enterprise#10536

Related: odoo/odoo#50284
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>