[ADD] mail_enterprise: include maintenance info in update data
The maintenance data contains limited information about the quantity of
custom code being used on the database, either via non-standard module
("extra modules"), or via "server actions" or "computed fields" declared
directly inside the database (e.g. with Odoo Studio).

If the customer chooses to buy the Maintenance service, this information
will ensure that all customizations are accounted for and covered for
all the relevant services (bug fixes, upgrade, etc.)

Users who decide to not buy the Maintenance service and are concerned
about this limited info[1] being sent, can opt-out by creating a key in
the System Parameters of the database:
`publisher_warranty.maintenance_disable` (value can be anything, e.g. true)
This may hamper the ability to obtain the service in the future, of
course.

[1] A global number of lines of code per "extra module", or possibly an
error status when parsing fails - cfr odoo/odoo#52016 for the details.

closes odoo/enterprise#11018

X-original-commit: 14c6b6685bd04c4bfe83608f7ca7b34fa75ce2c2
Related: odoo/odoo#52635
Signed-off-by: Olivier Dony (odo) <odo@openerp.com>