[FIX] marketing_automation: fix campaign duplication and participant view
Purpose
=======

When duplicating a campaign we should not duplicate require_sync and
traces from child activities.

We also fix the participant view, when seeing marketing traces. The
"Not bounced yet" message is not displayed, leading to a strange blank
block. We add missing text and its translation as it is more a fix than
a display improvement.

Task ID-2243914

closes odoo/enterprise#13707

X-original-commit: 96bdedd14fbd003db5bbf026b7c334156403c054
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>