[FIX] timesheet_grid: Fix the timer visibility in timesheet grid view
Currently when manualy encode an entry in the future then launch the
timer, the timer appears not to be launching, but it is still running
in the list view. Then, the timer cannot be launched for any other
project/task unless the user refreshes the page.

So after this commit, the timer will appear when the timer is started,
even if the timesheet grid has a manually encoded entry before future
date.
The issue is produced because from _searchTimesheet it was always
return first record even there is multiple record exist. it was not
checking on all the record for the today date and adjustment. so check
all the record and return the list for record instead of single record.

closes odoo/enterprise#10750

Taskid: 2243313
Closes: #10286
X-original-commit: dd83844
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>