[FIX] web_studio: race condition in test tour
Adding the second rule will trigger a complete redraw of the sidebar's
content, which might make the content of the selector for the first
domain button stale by the time it tries to run.

Instead of waiting for the snack bar to be done, wait for a second
approval rule to be present in the dom to ensure that this step only
runs once the sidebar's dom is completely redrawn.

closes odoo/enterprise#10773

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>