[FIX] web_studio: allow approval flows for 'type="action"' buttons
Due to a DistractedDevelopperError, the click handler for buttons in the
header bar was only added if the button was of type 'object', but this
restriction is no longer necessary - and in fact prevents the user from
setting approval flows on button that trigger actions through their id.

Since the sidebar now knows whether it needs to display the rainbow man
configuration section based on the button's spec, this condition was no
longer necessary and all clicks on header buttons should trigger a
sidebar redraw.