[FIX]web_studio: show widget when description is not set
When using web studio, whenever a technical widget is already set on a field and
the user customizes it with studio out of the debug mode, the 'widget' option
will show him an erroneous information, by simply fallbacking on the first value
available.

By this commit, this issue has been fixed.

Task - 2248455

closes odoo/enterprise#10866

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>
Co-authored-by: Mitali Patel <mpa@odoo.com>