[IMP] account: tax report: add the possibility to use a different country than the company's for the tax report
Some companies have their office in a country, and sell and pay taxes in another one. Before this commit, they couldn't use Odoo's tax report, as the country set on the res.company than the one to be used with their taxes. It is now possible to do that, using a config parameter.

closes odoo/enterprise#11039

X-original-commit: 414a633318fa1cf44c07b9caff34fb6445706ae3
Related: odoo/odoo#52734
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>