[IMP] iot: Download drivers from installed modules only.
Previously, all drivers and interfaces found on the server were
downloaded to the IoT Box. Having unused interfaces and drivers on the
IoT Box slowed it down as threads were running for nothing.

It was also confusing for some users to see those drivers on the IoT
Box homepage (e.g. SerialBlackBoxDriver when located outside of
Belgium)

We now download only the handlers from installed modules + the base
handlers located in hw_drivers.

closes odoo/enterprise#10818

Taskid: 2195956
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>