[FIX] hr_contract_salary: Flush contract simulation as sudo
We flush the contract to avoid writing anything after the rollback (maybe
everything should be flushed and not only the contract?).
Flushing triggers fields recomputation which read and writes some records.
However the user might not have the required access rights here.

This leads to the following problem:
Install `test_l10n_be_hr_payroll_account` only and run the module tests.
=> The tour fails, trying to read the employee without the required access
rights.

Note: the tour does not fail if all modules are installed; hence it does not
fail when tests are run on Runbot. The reason is that the employee is already
loaded (by some other module) in the cache with a sudo'ed environnement.

closes odoo/enterprise#10810

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>