[FIX] account: tax report: cash basis taxes: also consider cash payments
Before that, journals of type 'cash' were not excluded from the tax report SQL query, and caused some line duplications with the join, leading to incorrect amounts. We want to exclude all payments to only consider the invoice in that part of the query, so cash payments as well. We rephrase that here to only include sale and purchase journals (so that reconciling an invoice wil a misc entry instead of a payment will work as well)

closes odoo/enterprise#10927

X-original-commit: dafd4eb5379a8287c5157dd6fd3fb5abfb6a3886
Related: odoo/odoo#52381
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>