[IMP] web_enterprise: tooltip remain on top of other elements
Before this commit, the tooltip was shown when a element has focus or
the cursor was over. But on mobile the cursor move on tap event only and
so the tooltip disappear only when the user tap somewhere else.

After this commit, the tooltip appears only when the user touch the
element.

Steps to reproduce:
* Open Odoo on Mobile
* Go to CRM app
* Select a lead
* Go in edit mode
* Tap on the "Expected Closing" label field (BUG)
As you can see the tootip is over the datepicker.
Which is a bit annoying to select a date.

Task ID: 2152124

closes odoo/enterprise#10864

Related: odoo/odoo#52170
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>