[FIX] sale_coupon_taxcloud: prevent modification of invoiced sol
using Tax Cloud, create the invoice from the
SO and once the invoice is paid, add a free product on the SO.

When the user wants to create the second invoice (in order to make the SO 'fully
invoiced') an error message raise.

To avoid such situations we prevent the user to add products
when using Tax Cloud and promotion program lines with a UserError message.
The user should refund all existing invoices for this SO
(so that any 'invoiced' line get back to being uninvoiced),
then re-issue an invoice for the whole order.

opw-2261069

closes odoo/enterprise#10955

X-original-commit: 3651e06d1bece09f1a9e5572a4a5f693de367c8b
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>