[IMP] various: Remove depricated activity_exception widget from tree view
Purpose of the task is to remove the activity_exception widgets to
avoid redundancy. We have already added the activity on some list view
so activity expection widget on those list view is useless.

So in this commit, remove the activity_exception widget from
following tree view
 - helpdesk.ticket.tree
 - sale.subscription.list

closes odoo/enterprise#10846

Taskid: 2266720
Closes: #10846
Related: odoo/odoo#52109
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>