[FIX] l10n_mx_edi: Cancel process was improved
On v13 the invoice and payments could be reset2draft without cancel the record.

In Mexico, when a fiscal document is cancelled must be send the request to the SAT
to allow cancel it.

Now, when reset2draft not must e clear the stamp values if the record state is not cancelled,
this to avoid inconsistency between Odoo and the SAT.

Extra: The method ``_l10n_mx_edi_update_hour_timezone`` only is called when the time invoice
is not set, to avoid stamp 2 times a same invoice in the next case:
1. Create and validate the customer bill
2. Reset 2 draft
3. Validate again.

In 3, if the time invoice is different generate a new fiscal folio and is incorrect.

closes odoo/enterprise#10988

X-original-commit: a5e340aee9f210bfe3fd2a01abc03aa8913a303c
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>