[IMP] documents_spreadsheet: Add a menu to re-insert a pivot
Purpose
=======
Allow the user to easily 'regenerate' an outdated pivot by
simply re-inserting its updated version in the sheet.

Specification
=============
'Re-insert pivot' option in the context menu of cells.
This option unfolds on a list of all the pivots currently in the spreadsheet.
The 'display name' of a pivot is the model name followed by (#1), (#2), etc.
The selected pivot is inserted in its entirety from the active cell.

For this V1, we don't mind if the newly inserted pivot erase some
existing cells. The user can CTRL+Z if needed.

Task 2250129