[IMP] stock_barcode_picking_batch: add # lines to kanban
This commit adds the number of move lines to the kanban view of batch
stock pickings in barcode app. Note that this reflects the number of
move lines (i.e. the number of lines that show in the barcode app when
clicking on a batch record), not the number of moves (i.e. 0 done/
reserved products).

Part of "1. Improve Batch Pickings" specification of overall barcode app
improvements task.

Task: 1884520
Related Community PR: odoo/odoo#55096
Closes: odoo/enterprise#12086