[FIX] planning: do not force AM/PM time on slot template
Slot templates are localized, but we force the use of AM/PM which
depending on locale may be unexpected.

With this change, we use the "short" time format and to keep it small
for much timezone still replace ":00 " by space (this is only done for
AM/PM format since there is no space after minutes in 24 hours format).

opw-2266483
closes #10883

closes odoo/enterprise#10893

X-original-commit: d98e82a263369056b049111b41e7a0039862eac5
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>