[FIX] planning: Don't base default hours on calendar if different timezones
Purpose
=======
If a user in timezone X creates a planning slot for employee in timezone Y
(from the gantt view): the default hours are midnight to midnight (in the
user's timezone). Which are then adjusted to the employee's calendar.
However, the period midnight to midnight spans across two days in the
employee's timezone. This leads to weird situations when searching for
the closest attendance in the calendar.

E.g.
User in timezone "Asia/Colombo" (+05:30) and employee in timezone
"Europe/Brussels" (+02:00)
with a standard 40h/week calendar (from 8am to 12 and 1pm to 5pm everyday)
Create a slot the 2nd June.
Default hours are 00:00 to 23:59 (in "Asia/Colombo").
Which is 1st June 18:30 to 2nd June 18:29 (in "UTC")
Trying to find the closest attendances in the employee's calendar would
give 1st June 1pm to 2nd June 5pm (in "UTC").
Which displays 1st June 4:30pm to 2nd June 10:30pm (in "Asia/Colombo") to the
user who might not understand where those times are coming from.

Specification
=============
Do not try to use the employee's calendar if the user and employee have
different timezones.

closes odoo/enterprise#11399

X-original-commit: e7f3b307248b8a4d78db2fcde6519e9322519d6b
Related: odoo/odoo#53582
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>