[IMP] planning: Improve the planning analysis report
Currently, there is no distinction between the gantt view in the
reporting and in the other menus. All of the features of Planning
(shift creation, duplication, sending...) are still available,
which doesn't make sense for the reporting.

The planing analaysis report in palnning is based on seprate model
which just get data from planning.slot and display it but now we wants
report gantt view same as slot gantt view so remove the report model
and directly display the report from plannning.slot

So in this commit, Improve the below point:
 - remove the 'Planning Analysis' report from Project
  (keep it in the Planning app)
 - add the gantt and list views to the 'Planning Analysis' report
  (in planning app)
 - make the gantt view read-only in the reporting
 - Hide the copy previous week, send schedule and add buttons

closes odoo/enterprise#10943

Taskid: 2264393
Closes: #10943
Related: odoo/upgrade#1377
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>