[IMP] sale_subscription: Modify payment method on template
Currently, "Sen and try to charge" will send one extra e-mail during
the processing - if there's a delay between the creation of the invoice
and the actual processing of the payment (e.g. Sepa Direct Debit), then
the customer will at least be made aware of the invoice in advance.

But for SDD we warn the customer through the mandate and they
receive the invoice once it's paid. They can see what they're paying
from the portal... Also, the email is rather misleading as it seems
to imply that the customer need to do something So Removing that
payment mode would make sense, since the naming is not even clear.

So in this commit improved the payment method of subscription template
by firstly renaming the "Send" option by "Post" because when odoo
creates an invoice, it is not send but "posted".
And secondly remove "send & try to charge" option because the behavior
of this option is similar to "Send after successful payment" option.

closes odoo/enterprise#10942

Taskid: 2201199
Closes: #10942
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>