[IMP] hr_referral: add a total of point to be spend on dashboard
Purpose of the task is we never know how many points to get buy a
reward because remaining points to spend is not shown in 'hr.referral'.

So in this commit, you can see how many 'remaining points to be spend'
remain to get the reward and Also improve the design on dashboard and
move the number of level under the picture & total point on left and
point to spend on right.

For the admin hired application are create in recruitment and directly
in the stage contract signed so admin will not get the referral point
so in hr_referral make reward point for admin for demo data.

closes odoo/enterprise#11004

Task-id: 2227757
Closes: #11004
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>