[FIX] social_sale: allow users to see posted material
1. Install social_sale, website
2. Prepare a user with social media manager permissions and nothing else
3. Navigate to "Social Marketing > Posts" and create a post
4. Go to Posts again

Access error will popup because of computed value from resticted models
(sale_order). The user should be able to view the numbers without
accessing the sensitive data.
Removing group restrictions and allowing the computation to be
done in sudo mode allow the value to be computed correclty but the
permissions on the models are respected so the user is not able to
access quotations/sales view via smart button if he/she don't have the
rights to do so.

opw-2263242

closes odoo/enterprise#11001

X-original-commit: aae7cbd5890b667cfe9e6c3faa1f78615a3cb6a9
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>