[FIX] l10n_de_reports: correct letter label for 0% invoices
Create an invoice, assign a customer and create a single invoice
line with product and 0% tax.
Go to Accounting>General Ledger and click on "Export Datev (ZIP)".
The file EXTF_accounting_entries.csv into the zip package is will
report the acccounting lines.

The line related to the 0% invoice is marked as "s" instead of "h".

opw-2230699

closes odoo/enterprise#10970

X-original-commit: c43878ca32d178da43237330b1101c17a3ff9f6b
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>