[FIX] industry_fsm_sale_report: fix worksheet report generation - Install industry_fsm_sale_report - Go to Field Service > Configuration > Settings and activate "Worksheet Templates" - Go to Sales > Configuration > Settings and activate "Lock Confirmed Sales" - Create a new product:   * Product type: Service   In Invoice section of Sales tab:   * Service Tracking: Create a task in an existing project   * Project: Field Service - Create a new Quotation and add the created Service product - Confirm the quotation - Click on Tasks smart button - Click on Worksheet smart button - Write a comment and save - Go back to SO - Click on Tasks smart button - The "SIGN REPORT" and "SEND REPORT" actions are now available - Click on one of them The following error message appears : "It is forbidden to modify the following fields in a locked order: Quantity"
As introduced by this commit https://github.com/odoo/enterprise/commit/d04b48b5a27d30e63149ebf6783641a07c5d65ea,
when generating the Worksheet report, if the SO line is a task with timesheets, the value of the ordered quantity is updated with the value of
the delivered quantity, which generates an error because the SO is locked.
This has been done because the Worksheet report only displays ordered quantities but should display delivered quantities
for timesheets on tasks.

This is a weird behavior because generating a report should not modify the SO. That is why this behavior has been removed.
Instead of updating the SO lines, the report template has been modified to display delivered quantities
when the SO line is a task with timesheets and ordered quantities otherwise.

opw-2242790

closes odoo/enterprise#11173

X-original-commit: 8feb6a38082da7afbe83c7ebe3c759d2eac8042e
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>