[FIX] hr_payroll: Fix clickall test timeout
Purpose of the task is when all the demo data installed, the loading of
the page could take more than 5 seconds, as the first loading loads all
the work entries for all the employees on the given period. here, we
generate several hundreds records.

So in this commit, Load this month and the last month work entries,
to avoid creating all those records when running the click all.

closes odoo/enterprise#11228

Taskid: 2228945
Closes: #11019
X-original-commit: d47f4f3c674c85a8c181a33e56d03d61a5c8eb22
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>