[FIX] iot: instantiate service with env
PR odoo/odoo#52097 changed the API of services, as they now take
the env instead of a parent at initialization. This commit adapts
the code in iot accordingly.

Note that this is a quick fix. It would be better not to manually
instantiate the service, as it is already automatically
instantiated since we register it to the service registry. A proper
fix will come later.