[ADD] documents_spreadsheet: o-spreadsheet integration
Integrate o-spreadsheet with Odoo:
- Create a spreadsheet from Document
- Insert a pivot in a spreadsheet

The spreadsheet library itself is added as a src and not a lib,
so odoo can detect its strings to be translated

A new action has been created to open a spreadsheet in Odoo.

To extend o-spreadsheet, we created a PivotPlugin, which manages the
storage and the managing of pivot objects in spreadsheet.

Task-id 2223145, 2258512

Co-authored-by: Florent Lejoly <fle@odoo.com>
Co-authored-by: LucasLefevre <lul@odoo.com>
Co-authored-by: Pierre Rousseau <pro@odoo.com>