[IMP] documents_spreadsheet: add autofill on pivot
Purpose
=======

Provide the user a way to update the 'structure' of his pivot

Specifications
==============

Autofill from a Pivot cell will insert the next cells of the pivot
in its real state (i.e. with the new values for dimensions, ...).
In case of autofill with one level of group-by and if this level is a
date, the autofill will simply increment the date (it will allow to
generate formulas for the next periods)

Task-ID 2228946