[IMP] documents_spreadsheet: Add sheet name and breadcrumb
Purpose
=======
Allow to see and edit the spreadsheet name from within
the spreadsheet itself.

Specification
=============
Adds a breadcrumb with previous actions.
The last element in the breadcrumb should be
the speadsheet name. The name should be editable "inplace"
and should implement the same behavior as the name input
in Google Sheets.

Task 2238915