[FIX] l10n_de_reports: correct letter label for 0% credit notes
Create an invoice, assign a customer and create a single credit
note with product and 0% tax.
Go to Accounting>General Ledger and click on "Export Datev (ZIP)".
The file EXTF_accounting_entries.csv into the zip package is will
report the acccounting lines.

The line related to the 0% invoice is marked as "h" instead of "s".

opw-2230699

closes odoo/enterprise#11055

X-original-commit: f8970fc826242cb5b5efbf77bf6f4df48a712148
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>