[IMP] account_auto_transfer: add partner condition
Adds the ability to apply the Automatic Transfer only on specific
partners.

Also allows to have multiple filtered lines on an Automatic Transfer.
Also allows to have mix filtered and percent lines
These are already implemented features that where not allowed due to a
misscalculation of the total percentage of the lines

Adds tests for partner_id

Specify the fields that are displayed on account_ids (code, name,
user_type_id)

task-2247456

closes odoo/enterprise#10890

Related: odoo/upgrade#1638
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>