[IMP] Sign: Improve access rights by adding an employee group
PURPOSE
=======

The aims of this taks is to specified access of each templates of Sign.
Currently, we have to define a lot of people as "Administrator" of Sign app,
and this people have access to everything. With Template access group, we
can sort who can send a contract and  who have access to signed template.

SPECIFICATIONS
==============

In developer mode, properties of a template
- Change "Privacy" by "Who can Sign"
- Add field : "Template Access Group" => Only User on this group'll be
  able to have an access to this template. (Default : Empty (!! user who
  upload it has always access to it)

In developer mode, edit a template
- Add fields "Who can Sign" and "Template access group"

Change Acces right of Types of user:

Sign/Employee
- See sign in home
- Cannot upload a new templates but can upload a pdf to sign (oneshot)
- Can send a signature request through template he has rights on it
- See signed document he requested or he signed

Sign/User
- See sign menu in the app switcher
- can upload / send requests (template + pdf to sign)
- see templates he owned or he has right to see (privacy settings)
- see  signed documents of documents he signed or templates he owned or he requested

Sign /Administrator
- Access everything

closes odoo/enterprise#11126

Taskid: 2267471
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>