[FIX] web_grid: Fix alignment of text in grid view
Currently, when the encoding unit of timesheet is 'day', it will going
to display the toggle button for timesheet in grid view and the
aligment in button is center and there is padding for button so it is
not aligned same as unit hours.

So in this commit, Align the buttons text to the right and remove the
right padding in the grid view for encoding unit days.

closes odoo/enterprise#11225

Task-id: 2275993
Closes: #11225
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>