[IMP] delivery_barcode, delivery_{connectors}, stock_barcode(_picking_batch): add/update "Put In Pack"
This commit sets up the logic and adds the button for "Put in Pack"
in the barcode app for batch pickings. Also does corresponding renaming
of 'put_in_pack' method to 'action_put_in_pack' done in community to
make naming consistent with batch picking implementation.

This commit also adds tests for batch level "Put in Pack" button. Since
click behavior and end result expected to be the same as for a picking
in the barcode app, the same tours are re-used.

Part of "1. Improve Batch Pickings" specification of overall barcode
improvements task.

Task: 1884520
Community PR: odoo/odoo#55096
Closes: odoo/enterprise#12086