[FIX] web_studio: fix adding a field to a tree view inside a form view - Go to any form with a tree view inside it - For example, go to Sales and open a quotation - Toggle Studio - Click on "Order Lines" table and click on "EDIT LIST VIEW" button - Add a new field in the table An error is triggered:     Error while validating view:     Element '<xpath expr="//field[@name='order_line']/tree/tree[1]/field[7]">' cannot be located in parent view
Since this commit https://github.com/odoo/enterprise/commit/4c99531d6776ac9165a266b3d95890d1450edfa9,
absoluted xpath is prioritized, but in this case, a relative xpath has to be used because the xpath is computed
relative to the xpath of the subview.

A relative xpath will be used if "subview_xpath" is given.

opw-2276256

closes odoo/enterprise#11224

X-original-commit: 4dcf3f18c164dd6e6f6fb4460dcb68d9c35aa087
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>