[FIX] marketing_automation: fix performance in sync_participant
For campaign with a lot of existing record
The unicity check prefetch all the field of all the existing
record in memory to only read on field for the unique criteria.
It may cause memory error for campaign with a lot of record

Solution: Avoid prefetching all the field by reading
the only field we need. This also speedup that part
of the process

For 460 000 lead the memory usage change
from 2.5 GB to 650MB
the process took 65s and now 15s

At the end of the process record that
does not match the domain anymore are removed
This mean they are move to stage unlinked

The first issue: the process spend time
to remove participant already removed
and it does it one by one. It take a very
long time to finish and if the process is
restarted it will start over again.

Solution: Remove only record not removed yet
and remove them 1000 by 1000. It improve the speed
by ~1000 times

closes odoo/enterprise#11369

X-original-commit: bd3b8871f58d9c80a3094f670d06f300e9ebaa35
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>