[FIX] sale_subscription: prevent cron creating invoice...
...in case of the sale.subscription.template is set to Manually

Create a subscription having the create invoice set to Manually
shouldn't create an invoice when the sale_subscription is in
state running unless it has been specifically asked for
(using the create invoice button).

As the domain wasn't filtering on this parameter, the 'manually'
was managed as the 'draft_invoice'.

completing opw-1935178
completing commit-entreprise/599fe85

opw-2278589

closes odoo/enterprise#11313

X-original-commit: fb0b57fbb5bc98ac3bc6cf8f552fd62ff9016011
Signed-off-by: Simon Goffin (sig) <sig@openerp.com>