[FIX] ocn_client: Some people cannot be tagged
Before this commit, when a tracking value was changed and OCN was active,
this change was send to notify also on Mobile when needed.
But an issue was added in the message generator as the
'message.subtype_id.description' can be null.

After this commit, we check that it's value is not null before processing it.

Note: The Javascript/XML counterpart already check this possibility.

Steps to reproduce:
* Install OCN
* User A must have no language set
* Subtypes "discussion"/"note" have no description nor translation of the description
* Log in the mobile Android app with user B => he'll get an OCN token
* User A pings user B in a chatter => crash

opw-2279589

closes odoo/enterprise#11275

X-original-commit: 2c2ba854b0c425c760cd01085c95ade47f3bb751
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>