[IMP] stock_barcode: allow partner_id assignment in pickings
This commit makes it so the partner_id is always visible and editable
in the information screen of stock_picking and stock_picking_batch in
barcode app. In order to do this, a check needed to be added before
opening the corresponding ViewsWidget so only pickings' info are
editable and not batch pickings or inventories info views. Additionally
an extra check had to be added into _onReload to avoid trying to access
record data that does not exist (i.e. error).

Specification "5. Select a partner from the app" of overall barcode
improvements task.

Task: 1884520
Related Community PR: odoo/odoo#55096
Closes: odoo/enterprise#12086