[IMP] stock_barcode: add visual signal for extra products
This commit makes it so when a product that is not part of a picking's
initial move lines is scanned or manually +1 incremented, then its line
will be highlighted in red. To distinguish between when a picking has
initial move lines or not, the "immediate_transfer" field is checked.
This check also makes it so lines will not constantly turn red when
creating a new picking in the barcode app. Note that lines that are red
will stay red when a corresponding lot/sn is scanned and applied to it
(when doing the "scan product" => "scan lot/sn" flow). Red highlight
is not applied to Inventory Adjustments.

Specification "4. Visual Signal for wrong products" of overall barcode
improvements task.

Task: 1884520
Related Community PR: odoo/odoo#55096
Closes: odoo/enterprise#12086