[IMP] timesheet_grid: Revamp the timesheet kanban view with timer
Purpose of the task is to improve the usability of the timesheet
kanban view.

SPECIFICATIONS
==============

Kanban view revamp:
 - On the left: display the picture of the employee
 - In the middle: the name of the project and below, the name of the task
 - On the right: the duration with - / + buttons
When using the - / + buttons, the duration should be incremented of
x minutes based on the rounding number set in the settings of
Timesheets.
 - On mouse over on the duration, display a Play button
 - Clicking on the play button should launch the timer and on hover of
time dusplay the stop button  and clicking on the stop button
should effectively stop the timer and increment the duration
accordingly (no change of behavior).
 - The duration can still be edited while the timer is running by using
the - / + buttons and the timer should adapt accordingly.

closes odoo/enterprise#11497

Task-id: #2264613
Related: https://github.com/odoo/odoo/pull/53782
Related: odoo/odoo#53782
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>