[IMP] timesheet_grid: merge timesheet entries
Currently, when the timer is stopped, it either increments an existing
entry if there is no description or it creates a new one. The timer
tends to generate a lot of small entries. However, it is always best to
have one single timesheet entry that regroups everything done for a
given project on a specific day. This way, the customer doesn't know
what activity took what time exactly and cannot argue over small
details. In that context, a tool to merge all of the small entries
generated by the timer into a single one at the end of the day would be
a great help.

closes odoo/enterprise#11365

Taskid: 2275778
Related: odoo/odoo#53779
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>