[FIX] l10n_mx_edi: Partner view was improved.
The field for locality on the partner view has 2 options, char value or
a value from the SAT catalog, this second option is used when you need
to generate an CFDI with external trade complement, because needs a
value from the catalog.

Now, when you select the locality from the catalog hide the char field.
Same case that with the city and city_id here
https://github.com/odoo/odoo/blob/13.0/addons/base_address_city/models/res_partner.py#L37

closes odoo/enterprise#11451

X-original-commit: 6ca28ace04e3bfc5dde27ec12e1465cc84f44fc4
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>