[FIX] stock_barcode: fix alias detection
Have a product P recognizable by barcode B1
Create a product A recognizable by barcode B2
Add to the current barcode nomenclature an alias B2 -> B1 on top of
other rules
Create a sale order with delivery of P
Go to Barcode>Operations>Delivery Orders. Select the last delivery order
and scan B2.

A will be recognized and added to the delivery order,
despite the alias rule set. This occur because the function
that parses the barcode actually return its own input as long as the
parsing is successful.

opw-2280064

closes odoo/enterprise#11414

X-original-commit: 1f34f90ea7f240bcd3fcaa68b8ce1df4edd3d52c
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>