[FIX] hr_appraisal_survey: Fix traceback on ask feedback on appraisal.
Currently, When we ask a feedback (Survey integration) on an appraisal,
then the error is occurred due to key index which is not found in
default_get result.

So in this commit, we fix this issue by checking appraisal_id in result
if not found then set the active_id as appraisal_id.

closes odoo/enterprise#11524

Taskid: 2282089
Closes: #11396
X-original-commit: 1684c3461b7b4e15723675bb1c91ad6680994a79
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>