[FIX] hr_contract_salary: Improve the usability of salary configurator
Currently, In the contract form when user try to fill the details,
user need to remove 'Enter your name' text before add first name
and last namevand it also allow to submit the form without selection
of gender.

So in this commit, we fix this issue by adding placeholder in firstname
and lastname and making the employee name blank while creation
and make gender field required on salary configrator.

closes odoo/enterprise#11547

Taskid: 2267774
Closes: #11547
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>