[FIX] documents: cannot upload/replace files on iOS
When the user attempts to upload or replace a file in Documents from an
iOS device, the native file chooser opens but selecting a file doesn't
trigger the upload.

The reason behind this behaviour is that, on iOS, in order to have a
`onchange` event be triggered from a `<input type="file">` element, this
input have to be attached to the DOM.

This commit fixes it by actually attaching the hidden input to the DOM
and also prevents it from cluttering too much the DOM when the user
cancels its action (cf. cancel doesn't trigger any event).

opw-2260895

closes odoo/enterprise#11432

X-original-commit: c3c2b406ed69391be130173a33fc61177d5d377a
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>