[FIX] sale_amazon: use apparent encoding to decode Amazon responses
Amazon does not always replies with UTF-8 encoded responses, and the
specified encoding is sometimes wrong. This commit replaces the
response encoding by its apparent encoding to mitigate such cases.

In the process, it also makes the response text the unique acceptable
result data because the library was never capable of handling bytes.

opw-2267543

closes odoo/enterprise#11477

X-original-commit: f539c0d7bedd89a9f372a8ebeb5153f2da1919d4
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>