[IMP] documents: improvement usability of actions rules
Currently, 'Actions' page is before 'Conditions' page on the
Actions rules setting. That's not logical, before act something,
we have to define the conditions of this action.

So in this commit, we put 'Conditions' before 'Actions' on the
Action rules setting(workflow action form view). Also replace the
label of 'Rule Name' and 'Workspace' by 'Action Button Name' and
'Related Workspace' respectively.

closes odoo/enterprise#11436

Taskid: 2278812
Closes: #11436
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>