[IMP] sign: only display redirect button after signature if set
Purpose
=======

In case of the signed document is a part of a process (e.g. I have to sign
a document to have an access to another document) there is a "Redirect Link"
Feature.

The aims of this task is to force the signatory to follow the link implemented
on the "Redirect Link" field of the Template properties.

Currently, the Wizzard proposes differents options. If the Signatory don't click
on the right option, he doesn't have access to the redirect link.

Specification
=============

If the Redicrect Link field of the Template properties is filled
When the signatory finished to sign the Template, the Wizzard only proposes to go
to the Redirect Link

closes odoo/enterprise#9692

Taskid: 2230042
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>