Revert "[REM] iot_pairing: IoT Box Connection"
A skipci was made in order to merge a new module `iot_pairing` in saas-13.4.
Unfortunately after the module reached master, beb3851 was introduced before new saas-13.4 database
were generated, allowing to merge the change in master but introducing the failure in saas-13.4
for all other branches.
The failure should appear in master after saas-13.4 nightly.
This reverts commit beb3851b218f1f8f11f77f4a3afe9e4ff25f540b waiting for a fix nof upgrade scripts.

closes odoo/enterprise#11441

Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>